requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.59.0
django-braces==1.15.0
django-celery-results==2.5.1
django-colorfield==0.11.0
django-cors-headers==4.2.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
fcm-django==2.0.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.57.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
kombu==5.3.2
Markdown==3.4.3
msgpack==1.0.5
numpy==1.25.1
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
